Do not crash when loading save files with unknown content
This update fixes the crash that happens when you try to load a save exported from a
version of Idle Loops with unknown stats, skills, buffs, or options (like the prestige
fork or the dmchurch fork). It keeps the unknown entries in the save file to allow
two-way migration between forks without losing any (or at least, without losing much) data.